requirements.txt
mne
mne_bids
pandas
openpyxl
pyyaml
scipy
tqdm
pybv>=0.6